2_DSA_LLM/requirements.txt
langchain
langchainhub
langchain-core
langchain-groq
langchain-community
langgraph
google-search-results 
numexpr 
python-dotenv